Automatically generate tests for all combinations for int/real conversion
